requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
coverage==7.13.4
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
gunicorn==25.1.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
Werkzeug==3.1.6
